+++ src/strings/string_support.cpp
@@ -16,6 +16,9 @@
  *
  */
 
+#include <string.h>
+
+#include <extras/exceptions.hpp>
 #include <extras/strings/string_support.hpp>
 
 namespace extras {
@@ -53,6 +56,52 @@ namespace extras {
           to.length();  // Handles case where 'to' is a substring of 'from'
     }
     return str;
+  }
+
+  std::string replace_last(std::string str, std::string from, std::string to,
+                           char delim) {
+    StringContainsDelimException::assertion(str, delim, __INFO__);
+    if (!contains(str, from)) return str;
+    if (from.size() == 0) return str;
+    std::string dup = str;
+    std::string s_delim;
+    s_delim += delim;
+    dup = replace_all(dup, from, s_delim);
+    auto parts = split(dup, delim);
+    if (parts.size() == 1) {
+      return parts[0];
+    }
+    if (parts.size() == 2) {
+      // sweet jesus ...
+      auto dupa = str;
+      auto dupb = from;
+      reverse(dupa.begin(), dupa.end());
+      reverse(dupb.begin(), dupb.end());
+      if (strncmp(dupa.c_str(), dupb.c_str(), dupb.size()) == 0) {
+        dupa = dupa.substr(dupb.size());
+        reverse(dupa.begin(), dupa.end());
+        return dupa;
+      } else
+        return parts[0] + to + parts[1];
+    }
+    std::string rebuilt;
+    for (size_t i = 0; i < parts.size() - 2; i++) {
+      rebuilt += parts[i];
+      rebuilt += from;
+    }
+    rebuilt += parts[parts.size() - 2];
+    rebuilt += to;
+    rebuilt += parts[parts.size() - 1];
+    return rebuilt;
+  }
+
+  void StringContainsDelimException::assertion(const std::string &str,
+                                               char delim,
+                                               const extras::WhereAmI &ref) {
+    std::string test;
+    test += delim;
+    if (contains(str, test))
+      throw StringContainsDelimException(str, delim, ref);
   }
 
 }  // namespace extras